Support -u running as real user
This change allows you to pass a LOCAL_USER_ID environmental variable to the
Docker container. The container will ensure that the user id is a full user,
in that it has a home folder and an entry in /etc/passwd.

Additionally, this change fixes the issue where the pid1 process cannot be
sent a TERM signal, by consuming pid1 via the entrypoint.sh.

See https://denibertovic.com/posts/handling-permissions-with-docker-volumes/